third_party/blink/web_tests/svg/dynamic-updates/SVGPatternElement-svgdom-patternUnits-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'patternUnits' property of the SVGPatternElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS patternElement.patternUnits.baseVal is SVGUnitTypes.SVG_UNIT_TYPE_OBJECTBOUNDINGBOX
PASS patternElement.patternUnits.baseVal is SVGUnitTypes.SVG_UNIT_TYPE_USERSPACEONUSE
PASS successfullyParsed is true

TEST COMPLETE
